TP4 Correction et code plus propre

+++ src/fr/sdv/banque/entites/Credit.java
@@ -3,7 +3,7 @@ package fr.sdv.banque.entites;
 public class Credit extends Operation {
 
 
-    public Credit(int montant, String date) {
+    public Credit(float montant, String date) {
         super(montant, date);
     }
