+++ src/fr/sdv/banque/entites/Debit.java
@@ -3,7 +3,7 @@ package fr.sdv.banque.entites;
 public class Debit extends Operation{
 
 
-    public Debit(int montant, String date) {
+    public Debit(float montant, String date) {
         super(montant, date);
     }
